chore(docker): add basic Dockerfile
Builds the app in Docker so that the build and runtime
OS are the same. This is for dependencies with native
extensions to be compiled and executed on the same architecture.

The only drawback is that the Docker image gets to hold
the src files, but since this is a microservice boilerplate
the overhead in bytes should be minimal.